+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0
+
+- Change `to_link_header` to return a string ready to be used as header
+- Change `header_links` to actually return the links for the header in object form
+
 ## 0.2.0
 
 - Change `method_missing` and base methods for `Base`